requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
idna==3.4
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
urllib3==1.26.15
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dnspython==2.3.0
idna==3.4
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
urllib3==1.26.15
